Army Done
@Route("army") done all
Army can be trained and updates before rendering. It is done through a temp table that automatically deletes after a date is passed and is added to army count. All restrictions and errors are made (with messages) + a table for all trainings in progress with a queue of max 5 instances. Additional information about cost, availability and confirm page.