Battle Reports system implemented
Battle Reports are generated after battles and registrated like messages for each of the participants.
Added some more columns to Battles Entity to be more useful.
Added owner of report and duplicate the report for each user.
Two pages for all reports and a single chosen one with ability to remove one at a time.